implemented react router

+++ src/App.js
@@ -1,19 +1,25 @@
+import {BrowserRouter ,Routes, Route} from "react-router-dom";
+import {createContext} from "react";
+
 import AddCode from "./AddCode";
 import ListCodes from "./ListCodes";
 import Button from "./Button";
 import InputField from "./InputField";
 import { useEffect, useState } from "react";
-import Response from "./Response";
+import Request from "./Request";
 import RandomDigit from "./utils/RandomDigit";
 import EditCode from "./EditCode";
+import Layout from "./pages/Layout";
+import Sidebar from "./Sidebar";
 
+export const AppContext = createContext();
 export default function App() {
   const [items, setItems] = useState([]);
   const [newItem, setNewItem] = useState({});
   const [isAddFormOpen, setIsAddFormOpen] = useState(true);
   const [isEditFormOpen, setIsEditFormOpen] = useState(false);
 
-  const [response,setResponse] = useState("");
+  // const [response,setResponse] = useState("");
   const [userInput,setUserInput] = useState("");
   const [isResLoading,setIsResLoading] = useState(false);
   const [canDisplayInput,setCanDisplayInput] = useState(false);
@@ -23,27 +29,10 @@ export default function App() {
 
   const[itemsEdit,setItemsEdit] = useState({});
 
-  const key = "ussd_data";
-  function handleIsAddFormOpen() {
-    setIsAddFormOpen(!isAddFormOpen);
-  }
 
+  const key = "ussd_data";
   function handleOnItemChange(newItem) {
     setNewItem(newItem);
-  }
-
-  function handleUserInput (input) {
-    console.log('send button clicked');
-    setUserInput(input);
-  }
-
-  function handleSetResponse(data) {
-    setIsAddFormOpen(false);
-    if(data.MSGTYPE === false) {
-      setCanDisplayInput(false);
-    }
-    console.log(data)
-    setResponse(data);
   }
 
   function handleCodeClicked (request) {
@@ -51,17 +40,6 @@ export default function App() {
     console.log("url is being set");
     const sessionID = RandomDigit();
     setRequest({...request,sessionID});
-  }
-
-  function handleOnEditClicked(id) {
-    setIsAddFormOpen(false);
-    setIsEditFormOpen(true);
-    const copyItems = items;
-    const filtered = copyItems.find(item => {
-      if (item.id === id) {
-      return item;
-    }} );
-    setItemsEdit(filtered);
   }
 
   function handleOnEditItem (editedItem) {
@@ -77,15 +55,14 @@ export default function App() {
   }
 
   function handleOnDeleteItem (id) {
+    if (!window.confirm("Are you sure you want to delete ?")) {
+      return;
+    }
     let copyOfItems = items;
     const filtered = copyOfItems.filter(item=>item.id !== id);
     setItems(filtered);
     localStorage.setItem(key, JSON.stringify(filtered));
     setIsEditFormOpen(false);
-  }
-
-  function resetError() {
-    setError("");
   }
 
 
@@ -111,70 +88,44 @@ export default function App() {
   }, [newItem]);
 
 
-  useEffect(() => {
-    console.log('effect to make api call run');
-    if (request.url) {
-      const payload = {
-        USERID : "Spectrum",
-        MSISDN : request.phone,
-        SESSIONID : request.sessionID,
-        NETWORK : request.operator,
-        MSGTYPE : false,
-        USERDATA : userInput || request.ussd ,
-      };
-      console.log(payload);
-      setError("");
-      console.log("sendingg request");
-      setIsResLoading(true);
-      fetch(request.url, {
-        method: "POST",
-        headers: {
-          "Accept" : "application/json",
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify(payload)
-
-      })
-          .then((response) => {
-            if (!response.ok) {
-              console.log(response);
-              throw new Error("Request could not be processed");
-            }
-            return response.json()
-          })
-          .then(data=> {
-            handleSetResponse(data);
-            setIsResLoading(false);
-          })
-          .catch((error) => {
-            console.log(error);
-            setError(error);
-            setIsResLoading(false);
-          });
-    }
-  }, [request,userInput]);
-
-
   return (
-      <div className="content">
-        <div className="l-side">
-          <Button content="&#43; New Code" width={200} onClick={handleIsAddFormOpen} />
-          <ListCodes newItems={items} onCodeClicked={handleCodeClicked} onClickEdit={handleOnEditClicked} onClickDelete={handleOnDeleteItem}/>
-        </div>
-        <div className="r-side">
-          {isAddFormOpen ? (
-              <div className="form-container">
-                <AddCode onItemChange={handleOnItemChange} />
-              </div>
-          ) : null}
-          {isEditFormOpen ? (
-              <div className="form-container">
-                <EditCode item={itemsEdit} onEditItem={handleOnEditItem}></EditCode>
-              </div>
-          ): null}
-          {response && <Response response={response} isResLoading={isResLoading} error={error} resetError={resetError}/>}
-              {canDisplayInput && <InputField onSubmitInput={handleUserInput}/>}
-        </div>
-      </div>
+      <BrowserRouter>
+        <AppContext.Provider value={{
+          items : items,
+          handleOnDeleteItem,
+          canDisplayInput
+        }}
+        >
+        <Routes>
+          <Route index element={<Layout/>}/>
+          <Route path="app" element={<Layout/>}>
+            <Route path="create" element={<AddCode onItemChange={handleOnItemChange}/>}/>
+            <Route path=":id" element={<EditCode onEditItem={handleOnEditItem}/>}/>
+            <Route path="request" element={<Request items={items}/>}/>
+          </Route>
+          <Route path="*" element={<p>Page not found</p>}/>
+        </Routes>
+          </AppContext.Provider>
+      </BrowserRouter>
+      // <div className="content">
+      //   <div className="l-side">
+      //     <Button content="&#43; New Code" width={200} onClick={handleIsAddFormOpen} />
+      //     <ListCodes newItems={items} onCodeClicked={handleCodeClicked} onClickEdit={handleOnEditClicked} onClickDelete={handleOnDeleteItem}/>
+      //   </div>
+      //   <div className="r-side">
+      //     {isAddFormOpen ? (
+      //         <div className="form-container">
+      //           <AddCode onItemChange={handleOnItemChange} />
+      //         </div>
+      //     ) : null}
+      //     {isEditFormOpen ? (
+      //         <div className="form-container">
+      //           <EditCode item={itemsEdit} onEditItem={handleOnEditItem}></EditCode>
+      //         </div>
+      //     ): null}
+      //     {response && <Response response={response} isResLoading={isResLoading} error={error} resetError={resetError}/>}
+      //         {canDisplayInput && <InputField onSubmitInput={handleUserInput}/>}
+      //   </div>
+      // </div>
   );
 }
